refactor(solve): extract post-push GitHub sync from solve.auto-pr.lib.mjs
src/solve.auto-pr.lib.mjs was 1361 lines, over the 1350-line warning
threshold (#1593). The compare-API readiness poll and the branch
visibility check move to src/solve.auto-pr-push-sync.lib.mjs unchanged,
taking their collaborators as parameters.

tests/test-issue-1829-compare-api-transient.mjs asserts the delegation
chain at source level, so it now follows the poll into its new home.

solve.auto-pr.lib.mjs: 1361 -> 1225 lines.

Refs #2175
