+++ tests/test-issue-1829-compare-api-transient.mjs
@@ -138,13 +138,21 @@ console.log('\n📋 compare-readiness handler degrades gracefully on transient c
 
 const autoPrContent = execSync(`cat ${srcDir}/solve.auto-pr.lib.mjs`, { encoding: 'utf8' });
 const readinessContent = execSync(`cat ${srcDir}/solve.auto-pr-compare-readiness.lib.mjs`, { encoding: 'utf8' });
+// Issue #2175: the compare-API readiness poll itself moved one level further out,
+// into solve.auto-pr-push-sync.lib.mjs, to keep solve.auto-pr.lib.mjs under the
+// 1350-line warning threshold. The delegation chain asserted below is unchanged,
+// only its middle link now lives in that module.
+const pushSyncContent = execSync(`cat ${srcDir}/solve.auto-pr-push-sync.lib.mjs`, { encoding: 'utf8' });
 
 test('auto-pr lib imports and delegates to handleCompareApiNotReady', () => {
   if (!/import\s*\{[^}]*handleCompareApiNotReady[^}]*\}\s*from\s*'\.\/solve\.auto-pr-compare-readiness\.lib\.mjs'/.test(autoPrContent)) {
     throw new Error('solve.auto-pr.lib.mjs does not import handleCompareApiNotReady');
   }
-  if (!autoPrContent.includes('compareReady = await handleCompareApiNotReady(')) {
-    throw new Error('solve.auto-pr.lib.mjs does not delegate the not-ready decision to handleCompareApiNotReady');
+  if (!/waitForCompareApiReady\({[^}]*handleCompareApiNotReady/.test(autoPrContent)) {
+    throw new Error('solve.auto-pr.lib.mjs does not forward handleCompareApiNotReady into the compare-readiness poll');
+  }
+  if (!pushSyncContent.includes('compareReady = await handleCompareApiNotReady(')) {
+    throw new Error('solve.auto-pr-push-sync.lib.mjs does not delegate the not-ready decision to handleCompareApiNotReady');
   }
 });
